+++ src/meta/validation-error.ts
@@ -2,20 +2,25 @@ import Ajv from "ajv";
 import betterAjvErrors from "better-ajv-errors";
 import { MetaDataTable } from "./element";
 
-const metaSchema = require("../../elements/schema.json");
-
 export class MetaValidationError extends Error {
 	private obj: MetaDataTable;
+	private schema: any;
 	private errors: Ajv.ErrorObject[];
 
-	constructor(message: string, obj: MetaDataTable, errors: Ajv.ErrorObject[]) {
+	constructor(
+		message: string,
+		obj: MetaDataTable,
+		schema: any,
+		errors: Ajv.ErrorObject[]
+	) {
 		super(message);
 		this.obj = obj;
+		this.schema = schema;
 		this.errors = errors;
 	}
 
 	public prettyError(): void | betterAjvErrors.IOutputError[] {
-		return betterAjvErrors(metaSchema, this.obj, this.errors, {
+		return betterAjvErrors(this.schema, this.obj, this.errors, {
 			format: "cli",
 			indent: 2,
 		});